Move enum HolderState to protocols layer
