+++ libvcx/src/api_lib/api_handle/credential.rs
@@ -1,5 +1,5 @@
-use serde_json;
 use futures::FutureExt;
+use serde_json;
 
 use aries_vcx::agency_client::mocking::AgencyMockDecrypted;
 use aries_vcx::settings::indy_mocks_enabled;
@@ -284,13 +284,13 @@ pub async fn get_thread_id(handle: u32) -> VcxResult<String> {
 
 #[cfg(test)]
 pub mod tests {
+    use aries_vcx::protocols::issuance::holder::state_machine::HolderState;
     use aries_vcx::utils::devsetup::{SetupDefaults, SetupMocks};
     use aries_vcx::utils::mockdata::mockdata_credex::{ARIES_CREDENTIAL_OFFER, ARIES_CREDENTIAL_OFFER_JSON_FORMAT, ARIES_CREDENTIAL_RESPONSE, CREDENTIAL_SM_FINISHED};
     use aries_vcx::utils::mockdata::mockdata_credex;
 
     use crate::api_lib::api_handle::connection;
     use crate::api_lib::api_handle::credential::{credential_create_with_offer, get_attributes, get_credential, send_credential_request};
-    use crate::aries_vcx::handlers::issuance::holder::HolderState;
     use crate::aries_vcx::messages::issuance::credential::Credential;
     use crate::error::VcxErrorKind;
